Refactor feeds structure
